create super-class for datacontroller
