+++ src/dbt_magics/athenaMagics.py
@@ -1,4 +1,3 @@
-from time import time
 import pandas as pd
 from jinja2 import Template, Environment, meta
 import io
@@ -7,15 +6,96 @@ from pathlib import Path
 
 from IPython import get_ipython
 from IPython.core import magic_arguments, display
-from IPython.core.magic import line_magic, cell_magic, line_cell_magic, Magics, magics_class
-import ipywidgets as widgets
+from IPython.core.magic import line_cell_magic, Magics, magics_class
 
 import boto3
+from dbt_magics import datacontroller
 
 from dbt_magics.dbt_helper import dbtHelper, prStyle
 
+class AthenaDataContoller(datacontroller.DataController):
+    def __init__(self):
+        self.client = boto3.client('athena', region_name='eu-central-1')
 
-class Adapter(dbtHelper):
+        super().__init__()
+
+    """
+    Implemented Abstract methods
+    """
+
+    def set_dataset(self, database):
+        DatabaseList = self.list_databases(CatalogName=database)
+        self.wg_database.index = None
+        self.wg_database.options = tuple([i['Name'] for i in DatabaseList])
+    
+    def get_projects(self):
+        self.DataCatalogs = self.client.list_data_catalogs()['DataCatalogsSummary']
+        return [i['CatalogName'] for i in self.DataCatalogs]
+
+    
+    def set_tables(self, database):
+        self.TableMetadataList = self.list_table_metadata(CatalogName=self.wg_project.value, DatabaseName=database['new'])
+        self.wg_tables.index = None
+        self.wg_tables.options = tuple([i['Name'] for i in self.TableMetadataList])
+    
+    def set_columns(self, table):
+        f = lambda name, dtype: datacontroller.CB(name, dtype)
+        for i in self.TableMetadataList:
+            if i['Name']==table['new']:
+                cols = [(i['Name'], i['Type']) for i in i['Columns']]
+                # self.check_boxes = [f(f"{i['Name']} -- {i['Type']}") for i in i['Columns']]
+                self.partition_columns = [f(i['Name'], i['Type']+'(Part.)') for i in i.get('PartitionKeys', [])]
+                self.check_boxes = [f(i['Name'], i['Type']) for i in i['Columns']] + self.partition_columns
+                self.wg_check_boxes.children = self.check_boxes
+
+    """ 
+    Additional methods 
+    """
+
+    def list_databases(self, CatalogName):
+        response = self.client.list_databases(CatalogName=CatalogName, MaxResults=50)
+        DatabaseList = response['DatabaseList']
+        while 'NextToken' in response:
+            response = self.client.list_databases(CatalogName=CatalogName, MaxResults=50, NextToken=response["NextToken"])
+            DatabaseList += response['DatabaseList']        
+        return DatabaseList
+
+    def list_table_metadata(self, CatalogName, DatabaseName):
+        response = self.client.list_table_metadata(
+            CatalogName=CatalogName,
+            DatabaseName=DatabaseName,
+            MaxResults=50
+        )
+        TableMetadataList = response['TableMetadataList']
+        while 'NextToken' in response:
+            response = self.client.list_table_metadata(
+                            CatalogName=CatalogName,
+                            DatabaseName=DatabaseName,
+                            MaxResults=50,
+                            NextToken=response["NextToken"]
+            )
+            TableMetadataList += response['TableMetadataList']        
+        return TableMetadataList
+
+
+    def list_dataset_metadata(self, CatalogName, DatabaseName):
+        response = self.client.list_table_metadata(
+            CatalogName=CatalogName,
+            DatabaseName=DatabaseName,
+            MaxResults=50
+        )
+        TableMetadataList = response['TableMetadataList']
+        while 'NextToken' in response:
+            response = self.client.list_table_metadata(
+                            CatalogName=CatalogName,
+                            DatabaseName=DatabaseName,
+                            MaxResults=50,
+                            NextToken=response["NextToken"]
+            )
+            TableMetadataList += response['TableMetadataList']        
+        return TableMetadataList    
+
+class dbtHelperAdapter(dbtHelper):
     def __init__(self, profile_name="dbt_athena_dwh", target='prod'):
         super().__init__(profile_name=profile_name, target=target)
         self.DEFAULT_DBT_FOLDER = os.path.join(Path().home(), "projects", "data-aws", "dbt_athena_dwh")
@@ -118,11 +198,11 @@ SELECT * FROM {{ ref('table_in_dbt_project') }}
 ---------------------------------------------------------------------------
 """
         if cell is None:
-            dc = DataController()
+            dc = AthenaDataContoller()
             return dc()
         else:        
             args = magic_arguments.parse_argstring(self.athena, line)
-            self.dbt_helper = Adapter(args.profile, args.target)
+            self.dbt_helper = dbtHelperAdapter(args.profile, args.target)
             env = Environment()
             def ipython(variable):
                 try: result = get_ipython().ev(variable)
@@ -153,11 +233,6 @@ SELECT * FROM {{ ref('table_in_dbt_project') }}
                 df = df.head(int(args.n_output)) if type(df)==pd.DataFrame else None
                 return df
 
-
-
-
-
-
 def load_ipython_extension(ipython):
     js = """IPython.CodeCell.options_default.highlight_modes['magic_sql'] = {'reg':[/^%%(athena)/]};
     IPython.notebook.events.one('kernel_ready.Kernel', function(){
@@ -167,170 +242,3 @@ def load_ipython_extension(ipython):
     """
     display.display_javascript(js, raw=True)
     ipython.register_magics(SQLMagics)
-
-
-
-
-
-class prStyle():
-    BLACK = '\033[30m'
-    RED = '\033[31m'
-    GREEN = '\033[32m'
-    YELLOW = '\033[33m'
-    BLUE = '\033[34m'
-    MAGENTA = '\033[35m'
-    CYAN = '\033[36m'
-    WHITE = '\033[37m'
-    UNDERLINE = '\033[4m'
-    RESET = '\033[0m'
-    
-class CB(widgets.HBox):
-    def __init__(self, name, dtype):
-        super().__init__()
-        layout = widgets.Layout(max_width='450px')
-        self.add_class('aen_cb_style')
-        self.layout = layout
-        self.check = widgets.Checkbox(value=True, description=name, indent=False)
-        self.lab = widgets.Label(dtype.upper())
-        dtype = dtype if "int" not in dtype else "int"
-        dtype = 'decimal' if "decimal" in dtype else dtype
-        dtype = 'string' if "varchar" in dtype else dtype
-        self.lab.add_class(f'aen_cb_lab_style_{dtype}')
-        self.children = (self.check, self.lab)
-
-class DataController():
-    def __init__(self):
-        self.client = boto3.client('athena', region_name='eu-central-1')
-        self.DataCatalogs = self.client.list_data_catalogs()['DataCatalogsSummary']
-        self.catalog_names = [i['CatalogName'] for i in self.DataCatalogs]
-        
-        #----------- WIDGETS -----------#
-        self.wg_catalog = widgets.Dropdown(options=self.catalog_names, value=None)
-        self.wg_database = widgets.Dropdown(options=[])
-        self.wg_search_table = widgets.Text(placeholder='<table>')
-        self.wg_table = widgets.Dropdown(options=[])
-        self.wg_style = widgets.HTML('''<style>
-            .widget-text input[type="text"] {max-width:650px; background-color:#89d5c7; border-radius:7px; font-size: 13pt}
-            div.widget-dropdown select {max-width:650px; background-color:#73c6e3; border-radius:7px; font-size: 13pt} 
-            .wg_base_dropdowns {max-width:950px;}
-            .p-Accordion-child {border-radius:7px; max-width:650px;}
-            .widget-button {border-radius:7px; border: 2px solid green; font-family: "Monaco", monospace;}
-            .aen_cb_style {border-radius:7px; background-color:#c8d8dd; border: 1px solid black;}
-            .widget-label-basic span {font-size: 13pt;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_string {background-color:#89d5c7; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_date {background-color:#7d91e7; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_int {background-color:#c0d539; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_float {background-color:#e1a25a; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_boolean {background-color:#f59cf4; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_timestamp {background-color:#adbaf1; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_decimal {background-color:#e18972; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            .p-Widget.jupyter-widgets.widget-label.aen_cb_lab_style_double {background-color:#e1a25a; font-size: 19pt; border-radius:7px; text-align: center !important; color: black !important;}
-            </style>''')
-        self.wg_base_dropdowns = widgets.VBox(children=[self.wg_catalog, self.wg_database, self.wg_search_table, self.wg_table])
-        self.wg_base_dropdowns.add_class('wg_base_dropdowns')
-        self.all_columns = widgets.Button(description="All Columns")
-        self.wg_search_column = widgets.Text(placeholder='<column>')
-        self.wg_columns_and_search = widgets.HBox(children=[self.all_columns, self.wg_search_column])
-        self.wg_check_boxes = widgets.VBox(children=[])
-        self.wg_columns_container = widgets.VBox(children=[self.wg_columns_and_search, self.wg_check_boxes])
-        self.wg_column = widgets.Accordion(children=[self.wg_columns_container], selected_index=None)    
-        self.select_sql = widgets.Button(description="SELECT")
-        self.select_sql_star = widgets.Button(description="SELECT ALL")
-        self.select_box = widgets.VBox(children=[self.select_sql, self.select_sql_star])
-        self.output = widgets.Output()
-        self.output.add_class("anton-enns")
-        self.wg_output = widgets.Accordion(children=[self.output])
-        self.pannel = widgets.HBox(
-                children=[self.wg_base_dropdowns, self.wg_column, self.select_box, self.wg_output],
-                box_style='info' # one of 'success', 'info', 'warning' or 'danger', or ''
-            )  
-        
-        #----------- WIDGET ACTIONS -----------#
-        self.wg_catalog.observe(self.get_databases, names='value')
-        self.wg_database.observe(self.get_tables, names='value')
-        self.wg_table.observe(self.get_colums, names='value')
-        self.select_sql.on_click(self.on_button_clicked)
-        self.select_sql_star.on_click(lambda x: self.on_button_clicked(x, star=True))
-        self.wg_search_table.observe(self.search_tables, names='value')
-        self.all_columns.on_click(self.all_columns_handler)
-        self.wg_search_column.observe(self.search_columns, names='value')
-        
-        
-    def list_table_metadata(self, CatalogName, DatabaseName):
-        response = self.client.list_table_metadata(
-            CatalogName=CatalogName,
-            DatabaseName=DatabaseName,
-            MaxResults=50
-        )
-        TableMetadataList = response['TableMetadataList']
-        while 'NextToken' in response:
-            response = self.client.list_table_metadata(
-                            CatalogName=CatalogName,
-                            DatabaseName=DatabaseName,
-                            MaxResults=50,
-                            NextToken=response["NextToken"]
-            )
-            TableMetadataList += response['TableMetadataList']        
-        return TableMetadataList
-    
-    def list_databases(self, CatalogName):
-        response = self.client.list_databases(CatalogName=CatalogName, MaxResults=50)
-        DatabaseList = response['DatabaseList']
-        while 'NextToken' in response:
-            response = self.client.list_databases(CatalogName=CatalogName, MaxResults=50, NextToken=response["NextToken"])
-            DatabaseList += response['DatabaseList']        
-        return DatabaseList
-
-    def get_databases(self, catalog):
-        DatabaseList = self.list_databases(CatalogName=catalog['new'])
-        self.wg_database.index = None
-        self.wg_database.options = tuple([i['Name'] for i in DatabaseList])
-
-    def get_tables(self, database):
-        self.TableMetadataList = self.list_table_metadata(CatalogName=self.wg_catalog.value, DatabaseName=database['new'])
-        self.wg_table.index = None
-        self.wg_table.options = tuple([i['Name'] for i in self.TableMetadataList])
-
-    def get_colums(self, table):
-        f = lambda name, dtype: CB(name, dtype)
-        for i in self.TableMetadataList:
-            if i['Name']==table['new']:
-                cols = [(i['Name'], i['Type']) for i in i['Columns']]
-                # self.check_boxes = [f(f"{i['Name']} -- {i['Type']}") for i in i['Columns']]
-                self.partition_columns = [f(i['Name'], i['Type']+'(Part.)') for i in i.get('PartitionKeys', [])]
-                self.check_boxes = [f(i['Name'], i['Type']) for i in i['Columns']] + self.partition_columns
-                self.wg_check_boxes.children = self.check_boxes
-    
-    def on_button_clicked(self, x, star=False):
-        f = lambda name, dtype:  f'{name} {prStyle.BLUE}-- {dtype}{prStyle.RESET}'
-        with self.output:
-            self.output.clear_output()
-            part = [i.check.description for i in self.partition_columns if i.lab.value in ('DATE(PART.)', 'STRING(PART.)')]
-            if len(part):
-                part_string = f"\n{prStyle.MAGENTA}WHERE{prStyle.RESET} DATE({part[0]})=current_date\n{prStyle.MAGENTA}LIMIT{prStyle.RESET} {prStyle.CYAN}100{prStyle.RESET}"
-            else:
-                part_string = f"\n{prStyle.MAGENTA}LIMIT{prStyle.RESET} {prStyle.CYAN}100{prStyle.RESET}"
-            if star:
-                cols = '*'
-            else:
-                cols = "\n    , ".join([f(i.check.description, i.lab.value) for i in self.wg_check_boxes.children if i.check.value])
-            print(f'{prStyle.RED}%%athena{prStyle.RESET}\n{prStyle.MAGENTA}SELECT{prStyle.RESET}\n    {cols} \n{prStyle.MAGENTA}FROM{prStyle.RESET} {prStyle.GREEN}"{self.wg_catalog.value}"."{self.wg_database.value}"."{self.wg_table.value}"{prStyle.RESET}{part_string}')
-
-    def search_tables(self, observation):
-        self.wg_table.index = None
-        self.wg_table.options = tuple([i['Name'] for i in self.TableMetadataList if observation['new'] in i['Name']])
-
-    def search_columns(self, observation):
-        self.wg_check_boxes.children = [i for i in self.check_boxes if observation['new'] in i.check.description]
-
-    def all_columns_handler(self, observation):
-        if all([i.check.value for i in self.check_boxes]):
-            for box in self.check_boxes:
-                box.check.value = False
-        else:
-            for box in self.check_boxes:
-                box.check.value = True
-        # self.wg_column.children = [widgets.VBox([self.all_columns]+self.check_boxes)]
-    
-    def __call__(self):
-        return self.pannel
