+++ src/dbt_magics/magics.py
@@ -1,0 +1,7 @@
+"""
+Superclass for all dbt magics
+Includes generic GUI functions for all magics"""
+
+class Magics:
+    def __init__(self) -> None:
+        pass
